requirements.txt
Pillow>=8.0.0
numpy>=1.19.0
PyQt6>=6.0.0
psutil>=5.8.0
pynput>=1.7.0
